test: figure2d
